Kore.AST.PureML: Style guide conformance
