+++ src/main/haskell/kore/src/Kore/AST/PureToKore.hs
@@ -28,7 +28,6 @@ import Data.Functor.Impredicative
 import Kore.AST.Common
 import Kore.AST.Kore
 import Kore.AST.MetaOrObject
-import Kore.AST.PureML
 import Kore.AST.Sentence
 import Kore.Error
