+++ src/main/haskell/kore/test/Test/Kore/AST/PureToKore.hs
@@ -10,7 +10,6 @@ import Test.Tasty.QuickCheck
 import Kore.AST.Common
 import Kore.AST.Kore
 import Kore.AST.MetaOrObject
-import Kore.AST.PureML
 import Kore.AST.PureToKore
 import Kore.AST.Sentence
 import Kore.MetaML.AST
