+++ src/main/haskell/kore/test/Test/Kore/IndexedModule/MetadataTools.hs
@@ -17,7 +17,6 @@ import           Kore.AST.Builders
 import           Kore.AST.Common
 import           Kore.AST.Kore
 import           Kore.AST.MetaOrObject
-import           Kore.AST.PureML
 import           Kore.AST.PureToKore
 import           Kore.AST.Sentence
 import           Kore.ASTHelpers
